Add an exit 0 if no new tarball to do

+++ MLab/ShaperProbe/main.sh
@@ -39,6 +39,10 @@ sort -o done_tarballs.tmp done_tarballs.txt
 comm -23 latest_tarballs.tmp done_tarballs.tmp > ./tmp/tarballs_to_do.txt # delete lines which are in both files (and lines which are only in file done).
 rm -f latest_tarballs.tmp
 rm -f done_tarballs.tmp
+# if no new tarballs, exit
+if [ ! -s ./tmp/tarballs_to_do.txt ] ; then
+    exit 0
+fi
 # download the next tarball and process the current tarball at the same time
 line_old=$(head -n 1 ./tmp/tarballs_to_do.txt)
 echo $line_old | gsutil cp -I ./tmp/tarballs/ 2>> ./errors/download_tarballs.txt ; if [ $? -eq 0 ] ; then echo $line_old >> ./tmp/downloaded_tarballs.txt ; else rm -f tmp/tarballs/$(basename $line_old) ; fi # download the first tgz
@@ -67,8 +71,6 @@ echo "IP,date_test,server,clientversion,sleeptime,upshaper,minupburstsize,maxupb
 { echo ./csv/all/clean/*.csv | xargs cat; } >> data_raw.csv  # prevent the "Argument list too long" bug
 
 # POST-TREATEMENT
-
-# add an exit 0 if data_new.csv is empty 
 
 # set variables to be able to connect to mysql
 MYSQL_USER=$(sed -n -e 's/^MYSQL_USER="\([^"]*\)"$/\1/p' ../../Databases/mysql.conf)
